requirements.txt
# Kenya Startup Navigator - Backend Dependencies
# Windows-compatible version with simplified dependencies

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# AI/LLM Integration
groq==0.4.1

# HTTP Client for Groq API calls
httpx==0.25.2

# Data Validation and Settings
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Environment Configuration
python-dotenv==1.0.0

# Date/Time Handling
python-dateutil==2.8.2

# Security & Authentication (for future expansion)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Additional dependencies
typing-extensions==4.8.0
annotated-types==0.6.0